Move WordPressAuthenticator initialization to an extension for reusability
